runtime: Move runtime api code to the separate module
